+++ src/common/js/index.js
@@ -2,6 +2,6 @@
  * pwix:layout/src/common/js/index.js
  */
 
-import './config.js';
 import './constants.js';
-import './defaults.js';
+//
+import './configure.js';